moving plugins/core to plugins/common; moving filetype determination to common code
