+++ detect_secrets/plugins/common/initialize.py
