+++ tests/plugins/common/initialize_test.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 import mock
 import pytest
 
-from detect_secrets.plugins.core import initialize
+from detect_secrets.plugins.common import initialize
 from detect_secrets.plugins.high_entropy_strings import Base64HighEntropyString
 from detect_secrets.plugins.high_entropy_strings import HexHighEntropyString
