+++ tests/plugins/common/yaml_file_parser_test.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
-from detect_secrets.plugins.core.yaml_file_parser import YamlFileParser
+from detect_secrets.plugins.common.yaml_file_parser import YamlFileParser
 from testing.mocks import mock_file_object
